+++ poetry/masonry/builders/sdist.py
@@ -221,6 +221,9 @@ class SdistBuilder(Builder):
         for dep in sorted(self._meta.requires_dist):
             pkg_info += "Requires-Dist: {}\n".format(dep)
 
+        for name, url in sorted(self._meta.project_urls, key=lambda u: u[0]):
+            pkg_info += "Project-URL: {}, {}\n".format(name, url)
+
         return encode(pkg_info)
 
     def find_packages(self, include):