+++ tests/masonry/builders/test_complete.py
@@ -150,6 +150,8 @@ Provides-Extra: time
 Requires-Dist: cachy[msgpack] (>=0.2.0,<0.3.0)
 Requires-Dist: cleo (>=0.6,<0.7)
 Requires-Dist: pendulum (>=1.4,<2.0); extra == "time"
+Project-URL: Documentation, https://poetry.eustace.io/docs
+Project-URL: Repository, https://github.com/sdispater/poetry
 Description-Content-Type: text/x-rst
 
 My Package