+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added support for multi-constraints dependencies.
 - Added a cache version system.
 - Added a `--lock` option to `update` to only update the lock file without executing operations.
+- Added support for the `Project-URL` metadata.
 
 ### Changed